+++ offchain/inspect-server/src/main.rs
@@ -17,6 +17,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     env_logger::Builder::from_env(
         env_logger::Env::default().default_filter_or("info"),
     )
+    .target(env_logger::fmt::Target::Stdout)
     .init();
     let config = Config::initialize()?;
     log::info!("starting inspect server with {:?}", config);