+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Deploy to Arbitrum Goerli and Optimism Goerli
 - Add queue to serialize concurrent requests in inspect-server
 
+### Changed
+
+- Send inspect-server logs to stdout instead of stderr
+
 ## [0.5.0] 2022-08-17
 
 ### Added